[LWSD]: 完善初始化的处理

+++ src/exec/search/listend-ws/incl/lwsd_conf.h
@@ -26,7 +26,7 @@ typedef struct
 /* LWSD配置 */
 typedef struct
 {
-    int nid;                                /* 结点ID */
+    uint16_t nid;                           /* 结点ID */
     char name[NODE_MAX_LEN];                /* 结点名 */
     char wdir[FILE_PATH_MAX_LEN];           /* 工作路径 */
